added detection to light vs dark theme
