+++ src/App.jsx
@@ -1,7 +1,6 @@
 import React from "react";
 import { BrowserRouter as Router } from "react-router-dom";
 import Home from "./pages/home";
-import "./App.css";
 
 function App() {
   return (
