+++ src/Styles/Home.css
@@ -1,0 +1,148 @@
+/* General Styles */
+html {
+  scroll-behavior: smooth;
+  scroll-padding-top: 80px; 
+}
+
+* {
+  margin: 0;
+  padding: 0;
+  box-sizing: border-box;
+}
+
+body {
+  font-family: 'Inter', -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, sans-serif;
+  line-height: 1.6;
+}
+
+p{
+  color: #666;
+}
+
+.page-container {
+  min-height: 100vh;
+}
+
+.container {
+  max-width: 1200px;
+  margin: 0 auto;
+  padding: 0 20px;
+}
+
+section {
+  padding: 5rem 0;
+  scroll-margin-top: 80px; 
+}
+
+section h2 {
+  font-size: 2.5rem;
+  font-weight: bold;
+  text-align: center;
+  margin-bottom: 1rem;
+}
+
+.section-description {
+  text-align: center;
+  font-size: 1.2rem;
+  color: white;
+  max-width: 800px;
+  margin: 0 auto 3rem auto;
+}
+
+.main-header {
+  position: fixed;
+  width: 100%;
+  top: 0;
+  background-color: rgba(0, 0, 0, 0.5);
+  backdrop-filter: blur(8px);
+  z-index: 1000;
+}
+
+/* Responsive Design */
+@media (max-width: 768px) {
+  section {
+    padding: 3rem 0;
+  }
+
+  .hero-content h1 {
+    font-size: 2.5rem;
+  }
+  
+  .hero-subtitle {
+    font-size: 1.25rem;
+  }
+  
+  .hero-description {
+    font-size: 1rem;
+  }
+  
+  .nav-items {
+    gap: 1rem;
+  }
+  
+  .nav-items a {
+    font-size: 1rem;
+    padding: 0.25rem 0.5rem;
+  }
+  
+  .home-logo {
+    font-size: 20px;
+  }
+  
+  .social-links {
+    flex-direction: column;
+    gap: 1rem;
+    align-items: center;
+  }
+  
+  .social-link {
+    display: block;
+    width: 200px;
+  }
+}
+
+@media (prefers-color-scheme: dark) {
+  body {
+    background-color: #1f2937;
+    color: white;
+  }
+
+  .activities-section {
+    background-color: black;
+  }
+  .team-section{
+    background-color: black;
+  }
+
+  .hero {
+    background: linear-gradient(rgba(0, 0, 0, 0.6), rgba(0, 0, 0, 0.8)), 
+                url("https://images.unsplash.com/photo-1485827404703-89b55fcc595e?q=80&w=2070") center/cover no-repeat;
+  }
+
+  p{
+    color: white;
+  }
+
+  .activity-card {
+    background-color: #1f2937;
+    color: white;
+  }
+
+  .social-section {
+    background-color: #1f2937;
+  }
+
+  .social-link {
+    color: white;
+    border: 2px solid white;
+  }
+
+  .cta-button {
+    background-color: #3b82f6;
+    color: white;
+  }
+
+  .cta-button:hover {
+    background-color: #2563eb;
+  }
+}
