+++ src/Styles/NavBar.css
@@ -1,0 +1,42 @@
+nav {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    padding: 1rem 2rem;
+}
+  
+.home-logo {
+    font-size: 24px;
+    font-weight: bold;
+}
+  
+.home-logo a {
+    text-decoration: none;
+    color: white;
+    transition: color 0.3s ease;
+}
+ 
+.home-logo a:hover {
+    color: #3b82f6;
+}
+  
+.nav-items {
+    display: flex;
+    gap: 2rem;
+    list-style: none;
+}
+  
+.nav-items a {
+    text-decoration: none;
+    color: white;
+    font-size: 1.1rem;
+    font-weight: 500;
+    padding: 0.5rem 1rem;
+    border-radius: 0.5rem;
+    transition: all 0.3s ease;
+}
+  
+.nav-items a:hover {
+    background-color: rgba(255, 255, 255, 0.1);
+    color: #3b82f6;
+}
