+++ src/Styles/Team.css
@@ -1,0 +1,10 @@
+.team-section {
+    background-color: white;
+}
+  
+.team-grid {
+    display: grid;
+    grid-template-columns: repeat(auto-fit, minmax(280px, 1fr));
+    gap: 2rem;
+    padding: 0 1rem;
+}
