+++ src/Styles/TeamMember.css
@@ -1,0 +1,46 @@
+.team-member {
+    background-color: white;
+    border-radius: 1rem;
+    overflow: hidden;
+    box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
+    transition: transform 0.3s ease;
+}
+
+.team-member:hover {
+    transform: translateY(-5px);
+}
+
+.member-image {
+    width: 100%;
+    height: 300px;
+    overflow: hidden;
+}
+
+.member-image img {
+    width: 100%;
+    height: 100%;
+    object-fit: cover;
+}
+
+.member-info {
+    padding: 1.5rem;
+}
+
+.member-info h3 {
+    font-size: 1.5rem;
+    font-weight: bold;
+    margin-bottom: 0.5rem;
+}
+
+.member-info h4 {
+    color: #3b82f6;
+    font-size: 1.1rem;
+    margin-bottom: 1rem;
+}
+
+@media (prefers-color-scheme: dark){
+    .team-member {
+        background-color: #1f2937;
+        color: white;
+    }
+}
